+++ activerecord/CHANGELOG.md
@@ -1,3 +1,19 @@
+*   Add `set_constraints` helper for PostgreSQL
+
+    ```ruby
+    Post.create!(user_id: -1) # => ActiveRecord::InvalidForeignKey
+
+    Post.transaction do
+      Post.connection.set_constraints(:deferred)
+      p = Post.create!(user_id: -1)
+      u = User.create!
+      p.user = u
+      p.save!
+    end
+    ```
+
+    *Cody Cutrer*
+
 *   Include `ActiveModel::API` in `ActiveRecord::Base`
 
     *Sean Doyle*